refactor(sdk): simplify to single browser SDK
- Remove Node.js and local SDK variants
- Consolidate to single browser-optimized output
- Update generation scripts and workflows
- Simplify from 3 outputs to 1 for cleaner architecture
- Update CI to validate SDK stays current

Aligns with modern SDK practices where a single SDK handles
environment detection internally (Stripe, OpenAI, Supabase).
