+++ package.json
@@ -52,12 +52,9 @@
     "setup:dev": "pnpm install && echo '🎉 Development environment setup complete!'",
     "security:scan": "gitleaks detect --source=. --verbose --no-banner --no-git",
     "openapi:generate": "nx run-many -t openapi:generate",
-    "generate": "bash scripts/sdk-generate.sh",
+    "generate": "pnpm openapi:generate && fern generate --local",
     "sdk:generate": "bash scripts/sdk-generate.sh",
-    "sdk:check": "bash scripts/sdk-generate.sh --check",
     "fern:check": "fern check",
-    "fern:validate": "fern check && echo '✅ Fern configuration is valid'",
-    "prebuild": "pnpm generate",
     "prepare": "husky"
   },
   "keywords": [
