+++ packages/sdk/src/index.ts
@@ -14,10 +14,9 @@
 // Export core infrastructure
 export * from './core/index.js';
 
-// Export generated clients (when available)
-// These will be generated by Fern and placed in the generated/ directory
-// export * from './generated/browser/index.js';
-// export * from './generated/node/index.js';
+// Export generated client (when available)
+// This will be generated by Fern and placed in the generated/ directory
+// export * from './generated/index.js';
 
 // Export API functions
 export {
